implemented proxy caching

+++ crawler.py
@@ -43,13 +43,11 @@ def get_items_batch(executor: ProxiedRequestExecutor, after=None, url='https://o
 
 def generate_process_args(batch_count=20000, start=0, end=0, workers_count=1):
     '''bruh'''
-    proxies_page = 0
     for batch_start in range(start, end-1, -batch_count-1):
         batch_end = batch_start - batch_count
         batch_end = batch_end if batch_end > end else end
 
-        yield (batch_start, batch_end, proxies_page)
-        proxies_page = (proxies_page+1) % workers_count
+        yield (batch_start, batch_end)
 
 
 def maybe_save(markup, save_path, logger: logging.Logger):
@@ -86,12 +84,14 @@ def scrape_posts_batch(args: tuple):
     '''bruh'''
     dataset_root = './posts'
     dataset_root = './posts'
-    start, end, proxies_page = args
+    start, end = args
 
     logger = setup_logger(level=logging.INFO,
                           log_dir=os.path.join('.', 'logs', 'processes'))
     executor = ProxiedRequestExecutor(
-        logger=logger, proxies_page=proxies_page, max_sleep_time=0)
+        logger=logger,
+        max_sleep_time=0,
+    )
 
     files_saved = 0
 
@@ -123,18 +123,25 @@ def check_already_crawled(dataset_root):
     return len(batch), last_parsed_id
 
 
-def check_last_reddit_post_id():
+def check_last_reddit_post_id(last_if_not_found):
     '''bruh'''
     executor = ProxiedRequestExecutor(proxy_count=1)
     batch = get_items_batch(executor, url='https://old.reddit.com/new')
 
     last_post_id = get_id_from_thing(batch[0]).split('_')[-1]
-    last_id_if_none = np.base_repr(number=1847556708, base=36).lower()
+    last_id_if_none = np.base_repr(number=last_if_not_found, base=36).lower()
 
     return last_post_id if last_post_id is not None else last_id_if_none
 
 
-def scrape_reddit(dataset_root='./posts'):
+def is_available(url):
+    '''checks if website isnt down'''
+    executor = ProxiedRequestExecutor()
+    page = executor.get(url)
+    return page is not None
+
+
+def scrape_reddit(first_post_number: int, last_post_number: int, dataset_root='./posts'):
     '''scraping posts from old.reddit.com subreddits'''
 
     if not os.path.exists(dataset_root):
@@ -143,40 +150,46 @@ def scrape_reddit(dataset_root='./posts'):
     batch_count = 5000
     logger = setup_logger(filename='crawler.log')
 
-    reddit_first_post_ever = 295
+    reddit_first_post_ever = first_post_number
 
     crawled_count, last_crawled_id = check_already_crawled(dataset_root)
 
     logger.info('%d posts already crawled', crawled_count)
 
     if last_crawled_id is None:
-        last_crawled_id = check_last_reddit_post_id()
+        last_crawled_id = check_last_reddit_post_id(
+            last_if_not_found=last_post_number
+        )
         logger.info(
             'Starting crawling after hardcoded last post: %s', last_crawled_id)
     else:
         logger.info('Starting crawling after last saved post: %s',
                     last_crawled_id)
 
-    max_workers = min(4, (os.cpu_count() or 1))
-    with ProcessPoolExecutor(max_workers=max_workers) as pool:
-        try:
-            args_generator = generate_process_args(
-                batch_count=batch_count,
-                start=int(last_crawled_id, 36),
-                end=reddit_first_post_ever,
-                workers_count=max_workers,
-            )
-            futures_buffer = [pool.submit(scrape_posts_batch, args)
-                              for args in args_generator]
+    if is_available(url='https://old.reddit.com'):
+        max_workers = min(4, (os.cpu_count() or 1))
+        with ProcessPoolExecutor(max_workers=max_workers) as pool:
+            try:
+                args_generator = generate_process_args(
+                    batch_count=batch_count,
+                    start=int(last_crawled_id, 36),
+                    end=reddit_first_post_ever,
+                    workers_count=max_workers,
+                )
+                futures_buffer = [pool.submit(scrape_posts_batch, args)
+                                  for args in args_generator]
 
-            for future in as_completed(futures_buffer):
-                posts_saved = future.result()
-                logger.info('%d posts total saved from process', posts_saved)
+                for future in as_completed(futures_buffer):
+                    posts_saved = future.result()
+                    logger.info(
+                        '%d posts total saved from process', posts_saved)
 
-        except KeyboardInterrupt:
-            for pid in pool._processes:
-                os.kill(pid, signal.SIGKILL)
+            except KeyboardInterrupt:
+                for pid in pool._processes:
+                    os.kill(pid, signal.SIGKILL)
+    else:
+        logger.error('old.reddit is not available')
 
 
 if __name__ == '__main__':
-    scrape_reddit()
+    scrape_reddit(first_post_number=295, last_post_number=1847556708)
